Ironed out HTML/CSS
Configuring the toolbar, adding in a file icon on the left, page switching in the middle, and the other buttons on the right. Only the left and center got attention, and they still have errors. But lots of human written code, organized, and looks good on screen. Also added a new folder called icons which is where I call the icons from in the css.